extract `getCurrentHardfork`
